wip: test Windows build
