+++ pkg/thumbnail/thumbnail_windows.go
@@ -22,10 +22,3 @@ func BuildAnimatedThumbnail(inputBuf []byte, resp *http.Response) ([]byte, error
 	// Since the lilliput library currently does not support Windows, we error out early and fall back to the static thumbnail generation
 	return nil, errors.New("cannot build animated thumbnails on windows")
 }
-
-func BuildStaticThumbnail(inputBuf []byte, resp *http.Response) ([]byte, error) {
-	// govips can run on Windows with the proper setup. If you would like to contribute Windows
-	// support, see https://github.com/davidbyttow/govips#windows and open a PR at
-	// https://github.com/Chatterino/api/pulls.
-	return nil, errors.New("cannot build static thumbnails on windows")
-}
